chore: move nuxt image test to showcase example
