+++ examples/showcase/components/NuxtImage.stories.ts
@@ -1,15 +1,15 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 
-import MyComponent from '~/components/MyNuxtImage.vue'
+import NuxtImage from './NuxtImage.vue'
 
 // More on how to set up stories at: https://storybook.js.org/docs/vue/writing-stories/introduction
 
 const meta = {
-  title: 'Plugins/NuxtImage',
-  component: MyComponent,
+  title: 'Modules/Nuxt Image',
+  component: NuxtImage,
   // This component will have an automatically generated docsPage entry: https://storybook.js.org/docs/vue/writing-docs/autodocs
   tags: ['autodocs'],
-} satisfies Meta<typeof MyComponent>
+} satisfies Meta<typeof NuxtImage>
 
 export default meta
 type Story = StoryObj<typeof meta>
@@ -19,6 +19,6 @@ type Story = StoryObj<typeof meta>
  * to learn how to use render functions.
  */
 
-export const NuxtImage: Story = {
+export const NuxtImageStory: Story = {
   args: {},
 }
