+++ examples/showcase/components/NuxtImage.vue
